Update to Angular 17 (#228)
* Update ts version

* Update angular version

* Update clarity version

* Refactor markdown component

* Update cds core and add style sheets

* Add theme settings

* Update @angular/cdk

* Fix markdown rendering

* Set background color via variable to consider theme changes

* Implement custom text tooltip component

* Add prettier as dev dependency

* Add lint-staged as dev dependency

* Migrate code to new @if/@for/@switch syntax

* Update angular-split version

* Migrate ngx-dynamic-hooks to v3

* Replace deprecated xterm packages

* Fix path to xterm style sheet

* remove unused Web Components polyfills

* update @auth0/angular-jwt

* update ng2-charts

* update packages

* run nmp audit fix

* Add authn service. Improve theme handler

* Minor fix

* Ensure that note component always receives valid type

* Enable strict null checks & function types

* Fix bug that steps do not load initially

* Remove unused component

* Fix: Reference to scenario needed to correctly update steps

* Fix icons

* Remove unused code

* Replace clr-alerts by our alert component

* Minor fix

* Use "@else" or "@else if" where possible

* Simplify glossary component

* Fix nested md in hidden/node component

* Escape code to allow double quotes etc.

* Prevent removal of backslashes in ngx-dynamic-hooks inputs

* Clarity css vars changed ... add login padding

---------

+++ src/app/course/course.component.ts
@@ -8,6 +8,11 @@ import { DeleteConfirmationComponent } from '../delete-confirmation/delete-confi
 import { RbacService } from '../data/rbac.service';
 import { ClrDatagridSortOrder } from '@clr/angular';
 import { CourseWizardComponent } from './course-wizard/course-wizard.component';
+import { AlertComponent } from '../alert/alert.component';
+import {
+  DEFAULT_ALERT_ERROR_DURATION,
+  DEFAULT_ALERT_SUCCESS_DURATION,
+} from '../alert/alert';
 
 @Component({
   selector: 'app-course',
@@ -21,12 +26,9 @@ export class CourseComponent implements OnInit {
   @ViewChild('editcourse') editCourseWizard: CourseWizardComponent;
   @ViewChild('deleteConfirmation')
   deleteConfirmation: DeleteConfirmationComponent;
+  @ViewChild('alert') alert: AlertComponent;
 
-  public selectedCourse: Course;
-
-  public alertType: string = 'warning';
-  public alertText: string = '';
-  public alertClosed: boolean = true;
+  public selectedCourse: Course | null = null;
 
   public showActionOverflow: boolean = false;
 
@@ -39,9 +41,8 @@ export class CourseComponent implements OnInit {
   constructor(
     public courseService: CourseService,
     public scenarioService: ScenarioService,
-    public rbacService: RbacService
-  ) {
-  }
+    public rbacService: RbacService,
+  ) {}
 
   ngOnInit(): void {
     const authorizationRequests = Promise.all([
@@ -51,8 +52,8 @@ export class CourseComponent implements OnInit {
       this.rbacService.Grants('courses', 'delete'),
     ]);
 
-    authorizationRequests
-      .then((permissions: [boolean, boolean, boolean, boolean]) => {
+    authorizationRequests.then(
+      (permissions: [boolean, boolean, boolean, boolean]) => {
         const allowedGet: boolean = permissions[0];
         const allowedUpdate: boolean = permissions[1];
         const allowListScenarios: boolean = permissions[2];
@@ -64,29 +65,22 @@ export class CourseComponent implements OnInit {
         // Enable permission to update courses
         this.updateRbac = allowedUpdate && this.listScenarios;
         return this.listScenarios;
-      })
+      },
+    );
 
     this.refresh();
   }
 
   refresh(): void {
-    this.courseService.list(true).subscribe((cList: Course[]) => {
-      this.courses = cList;
+    this.courseService.list(true).subscribe({
+      next: (cList: Course[]) => {
+        this.courses = cList;
+      },
+      error: () => {
+        const alertMsg = 'Failed to load courses!';
+        this.alert.danger(alertMsg, true, DEFAULT_ALERT_ERROR_DURATION);
+      },
     });
-  }
-
-  alertSuccess(msg: string) {
-    this.alertType = 'success';
-    this.alertText = msg;
-    this.alertClosed = false;
-    setTimeout(() => (this.alertClosed = true), 1000);
-  }
-
-  alertDanger(msg: string) {
-    this.alertType = 'danger';
-    this.alertText = msg;
-    this.alertClosed = false;
-    setTimeout(() => (this.alertClosed = true), 3000);
   }
 
   openNewWizard() {
@@ -104,14 +98,21 @@ export class CourseComponent implements OnInit {
   }
 
   deleteSelected(): void {
+    if (!this.selectedCourse) {
+      const alertMsg = 'Error deleting object: Selected course is unavailable!';
+      this.alert.danger(alertMsg, true, DEFAULT_ALERT_ERROR_DURATION);
+      return;
+    }
     this.courseService.delete(this.selectedCourse).subscribe({
       next: (_s: ServerResponse) => {
-        this.alertSuccess('Course deleted');
+        const alertMsg = 'Course deleted!';
+        this.alert.success(alertMsg, true, DEFAULT_ALERT_SUCCESS_DURATION);
         this.refresh();
         this.selectedCourse = null;
       },
       error: (e: HttpErrorResponse) => {
-        this.alertDanger('Error deleting object: ' + e.error.message);
+        const alertMsg = 'Error deleting object: ' + e.error.message;
+        this.alert.danger(alertMsg, true, DEFAULT_ALERT_ERROR_DURATION);
       },
     });
   }
